Moved function `get_viewport_info` to SampleView component
